+++ modules/canvaskit/externs.js
@@ -468,6 +468,9 @@ var CanvasKit = {
     _addVerbsPointsWeights: function() {},
     _arc: function() {},
     _arcTo: function() {},
+    _arcToOval: function() {},
+    _arcToTangent: function() {},
+    _arcToRotated: function() {},
     _close: function() {},
     _conicTo: function() {},
     _cubicTo: function() {},
@@ -551,8 +554,9 @@ var CanvasKit = {
     makeImageSnapshot: function() {},
     makeSurface: function() {},
     sampleCnt: function() {},
-    reportBackendType: function() {},
+    reportBackendTypeIsGPU: function() {},
     grContext: {},
+    openGLversion: {},
 
     // private API
     _flush: function() {},
@@ -880,6 +884,9 @@ CanvasKit.SkPath.prototype.addRoundRect = function() {};
 CanvasKit.SkPath.prototype.addVerbsPointsWeights = function() {};
 CanvasKit.SkPath.prototype.arc = function() {};
 CanvasKit.SkPath.prototype.arcTo = function() {};
+CanvasKit.SkPath.prototype.arcToOval = function() {};
+CanvasKit.SkPath.prototype.arcToTangent = function() {};
+CanvasKit.SkPath.prototype.arcToRotated = function() {};
 CanvasKit.SkPath.prototype.close = function() {};
 CanvasKit.SkPath.prototype.conicTo = function() {};
 CanvasKit.SkPath.prototype.cubicTo = function() {};
@@ -1049,3 +1056,11 @@ var DOMMatrix = {
 
 // Not sure why this is needed - might be a bug in emsdk that this isn't properly declared.
 function loadWebAssemblyModule() {};
+
+// This is a part of emscripten's webgl glue code. Preserving this attribute is necessary
+// to override it in the puppeteer tests
+var LibraryEGL = {
+  contextAttributes: {
+    majorVersion: {}
+  }
+}