+++ src/gpu/GrProcessorUnitTest.h
@@ -52,12 +52,18 @@ std::unique_ptr<GrFragmentProcessor> MakeChildFP(GrProcessorTestData*);
 class GrProcessorTestData {
 public:
     using ViewInfo = std::tuple<GrSurfaceProxyView, GrColorType, SkAlphaType>;
-    GrProcessorTestData(SkRandom*, GrRecordingContext*, int numProxies, const ViewInfo[]);
+
+    GrProcessorTestData(SkRandom* random, GrRecordingContext* context,
+                        int numViews, const ViewInfo views[],
+                        std::unique_ptr<GrFragmentProcessor> inputFP = nullptr);
+    GrProcessorTestData(const GrProcessorTestData&) = delete;
+    ~GrProcessorTestData();
 
     GrRecordingContext* context() { return fContext; }
     GrProxyProvider* proxyProvider();
     const GrCaps* caps();
     SkArenaAlloc* allocator() { return fArena.get(); }
+    std::unique_ptr<GrFragmentProcessor> inputFP();
 
     ViewInfo randomView();
     ViewInfo randomAlphaOnlyView();
@@ -68,6 +74,7 @@ private:
     GrRecordingContext* fContext;
     SkTArray<ViewInfo> fViews;
     std::unique_ptr<SkArenaAlloc> fArena;
+    std::unique_ptr<GrFragmentProcessor> fInputFP;
 };
 
 class GrProcessor;